python-backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# CrewAI and AI framework
crewai==0.30.0
langchain>=0.1.10,<0.2.0
openai>=1.13.3,<2.0.0

# Composio integration
composio-core>=0.7.21
composio-langchain>=0.9.1

# Database and storage
sqlalchemy==2.0.23
alembic>=1.13.1,<2.0.0
redis==5.0.1

# API and web
httpx==0.25.2
websockets==12.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
asyncio-mqtt==0.13.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0